create folder of pages and move pages into respective path
